Adding the changes for likes and comments; correcting the data model again

+++ system/src/main/java/com/spotlight/system/service/model/FeedService.java
@@ -15,7 +15,7 @@ public class FeedService {
     }
 
     public Page<PostSummary> getFeed(int userId, Pageable pageable) {
-        return postRepository.getFeed(userId, pageable);
+        return postRepository.getFeed((long) userId, pageable);
     }
 
 }
